feat: empty frontend working app created and initialized

+++ backend/controllers/goalController.js
@@ -1,9 +1,11 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
+
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({user: req.user.id} );
 
   res.status(200).json(goals);
 });
@@ -15,6 +17,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user : req.user.id,
   });
 
   res.status(200).json(goal);
@@ -28,6 +31,16 @@ const updateGoals = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+const user = await User.findById(req.user.id);
+
+if(!user){ res.status(404);
+  throw new Error("User not found"); }
+
+
+  if (goal.user.toString()!== user.id) {
+    res.status(404)
+  throw new Error("User not authorized"); }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -40,10 +53,22 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  if(!user){ res.status(404);
+    throw new Error("User not found"); }
+  
+  
+    if (goal.user.toString()!== user.id) {
+      res.status(404)
+    throw new Error("User not authorized"); }
+
   await goal.deleteOne({_id: req.params.id});
 
   res.status(200).json({id: req.params.id});
 })
+
 
 module.exports = {
   getGoals,
